% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sscpls.R
\name{sscpls}
\alias{sscpls}
\title{Sparse Subspace Constrained Partial Least Squares}
\usage{
sscpls(X, Y, ncomp = 2, lambda, scale = F, center = T,
  compositional = F, abstol = 1e-06, reltol = 1e-06, max_itter = 10^5)
}
\arguments{
\item{X}{A matrix of regressors (n x p). By default the matrix will be
centered to have mean zero.}

\item{Y}{A matrix of continuous responses (n x q). By default the matrix will
be centered to have mean zero.}

\item{ncomp}{The number of components to include in the model.}

\item{lambda}{Thresholding parameter for sparsity}

\item{scale}{Scale predictors and response by dividing each variable by its sample standard deviation}

\item{center}{Center predictors and response by subtracting the column means from the variables}

\item{compositional}{Use either compositional projection or SIMPLS orthogonal projection}

\item{abstol}{Absolute tolerance for the ADMM algoithm (default is 1e-06)}

\item{reltol}{Relative tolerance for the ADMM algoithm (default is 1e-06)}

\item{max_itter}{Maximum number of itterations for the ADMM method}
}
\value{
\code{simpls} returns a fit PLS with:
\item{weights}{a list containing the X and Y pls weights.}
\item{scores}{a list containing the X and Y pls scores.}
\item{beta}{a regression coefficient}
}
\description{
Provides an implementation of an ADMM algorithm for computing the sparse
subspace constrained PLS estimator. Computes PLS regression where the
direction vectors are constrained to be both sparse and lie within a given subspace.
}
\examples{
n<-50 # number of observations
p<-50 # number of variables
X<-matrix(rnorm(n*p),ncol=p)
X <- scale(X)
y<-X[,1:5]\%*\%1:5 + matrix(rnorm(n))

res <- sscpls(X, y, ncomp = 5, lambda = 0.6, scale = F, abstol = 1e-07, reltol = 1e-07)

# abs and rel tol give a cutoff for the orthogonality constriant, i.e. here precision to 1e-06
round(crossprod(res$x.scores), 5)

}
